Remove IIIFv2 types/functions
