+++ catalogue/webapp/services/iiif/fetch/image.ts
@@ -1,9 +1,9 @@
 import { fetchJson } from '@weco/common/utils/http';
-import { Image } from '../types/image/v2';
+import { IIIFImage } from '../types/image/v2';
 
 export async function fetchIIIFImageJson(
   location: string
-): Promise<Image | undefined> {
+): Promise<IIIFImage | undefined> {
   try {
     const imageJson = await fetchJson(location);
     return imageJson;
