+++ catalogue/webapp/services/iiif/transformers/image.ts
@@ -1,6 +1,6 @@
-import { Image } from '../types/image/v2';
+import { IIIFImage } from '../types/image/v2';
 import { TransformedImageJSON } from '../../../types/image';
-export function transformImageJSON(imageJSON: Image): TransformedImageJSON {
+export function transformImageJSON(imageJSON: IIIFImage): TransformedImageJSON {
   return {
     width: imageJSON.width,
     height: imageJSON.height,
